+++ NEWS.md
@@ -18,7 +18,8 @@
   - Private methods have been unified so that they do not have the `RPolars` prefix (#584).
 
 ### What's changed
-
+- Use `pl$polars_code_completion()` to assist writing polars syntax. Support for native R terminal
+  and Rstudio. (#XYZ).
 - The Extract function (`[`) for DataFrame can use columns not included in the
   result for filtering (#547).
 - The Extract function (`[`) for LazyFrame can filter rows with Expressions (#547).
@@ -34,6 +35,7 @@
   precision (#589).
 - `<LazyFrame>$filter()`, `<DataFrame>$filter()`, and `pl$when()` now allow multiple conditions
   to be separated by commas, like `lf$filter(pl$col("foo") == 1, pl$col("bar") != 2)` (#598).
+
 
 ## polars 0.11.0